move settings screen into own component
